[REFACTOR] 패키지 위치 변경
기존: only domain패키지
변경: domain.entity, domain.document
