+++ src/main/java/com/example/whiplash/domain/entity/AssetType.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.domain;
+package com.example.whiplash.domain.entity;
 
 public enum AssetType {
     LARGE_CAP, MID_CAP, SMALL_CAP, PENNY_STOCK, HIGH_PRICE
