+++ src/main/java/com/example/whiplash/domain/entity/article/ArticleKeyword.java
@@ -1,0 +1,26 @@
+package com.example.whiplash.domain.entity.article;
+
+import com.example.whiplash.domain.entity.Keyword;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Table(name = "article_keywords")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder
+public class ArticleKeyword {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    // 단방향: ArticleKeyword → ArticleIndex
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "article_index_id", nullable = false)
+    private ArticleIndex articleIndex;
+
+    // 단방향: ArticleKeyword → Keyword
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "keyword_id", nullable = false)
+    private Keyword keyword;
+}
