+++ src/main/java/com/example/whiplash/domain/entity/history/search/SearchHistory.java
@@ -1,27 +1,26 @@
-package com.example.whiplash.domain;
+package com.example.whiplash.domain.entity.history.search;
 
+import com.example.whiplash.user.User;
 import jakarta.persistence.*;
 import lombok.*;
+import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "user_keywords")
+@Table(name = "search_history")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class UserKeyword {
+public class SearchHistory {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Integer priority;
+    private String keyword;
+    private Integer resultCount;
+    private LocalDateTime searchedAt;
 
-    // 단방향: UserKeyword → User
+    // 단방향: SearchHistory → User
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
-
-    // 단방향: UserKeyword → Keyword
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "keyword_id", nullable = false)
-    private Keyword keyword;
 }
