+++ src/main/java/com/example/whiplash/domain/entity/profile/ExperienceLevel.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.domain.profile;
+package com.example.whiplash.domain.entity.profile;
 
 public enum ExperienceLevel {
     BEGINNER, INTERMEDIATE, EXPERT
