+++ src/main/java/com/example/whiplash/domain/entity/profile/TradingStyle.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.domain.profile;
+package com.example.whiplash.domain.entity.profile;
 
 public enum TradingStyle {
     SCALPING, DAY_TRADING, SWING, POSITION
