+++ src/main/java/com/example/whiplash/domain/entity/trade/MockTradeLog.java
@@ -1,42 +1,41 @@
-package com.example.whiplash.domain.profile;
+package com.example.whiplash.domain.entity.trade;
 
 
-import com.example.whiplash.domain.user.User;
+import com.example.whiplash.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
+
 import static jakarta.persistence.EnumType.STRING;
 import static jakarta.persistence.FetchType.LAZY;
-import static jakarta.persistence.GenerationType.*;
+import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
 @Entity
-@Table(name = "investor_profiles")
+@Table(name = "mock_trade_logs")
 @Getter
 @NoArgsConstructor(access = PROTECTED)
 @AllArgsConstructor
 @Builder
-public class InvestorProfile {
+public class MockTradeLog {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
 
     @Enumerated(STRING)
-    private InvestorType investorType;
+    private TradeType tradeType;
 
-    @Enumerated(STRING)
-    private RiskTolerance riskTolerance;
+    private LocalDateTime tradeTime;
+    private Integer quantity;
+    private Integer pricePerUnit;
+    private Integer totalPrice;
+    private Integer fee;
 
-    private Integer investmentPeriod;
-    private Integer expectedReturn;
-
-    @Enumerated(value = STRING)
-    private ExperienceLevel experienceLevel;
-
-    // 단방향: Profile → User
+    // 단방향: Log → User
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
