+++ src/main/java/com/example/whiplash/domain/entity/trade/TradeType.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.domain.trade;
+package com.example.whiplash.domain.entity.trade;
 
 public enum TradeType {
     /**
