--- src/main/java/com/example/whiplash/domain/history/search/SearchHistory.java
@@ -1,26 +1,0 @@
-package com.example.whiplash.domain.history.search;
-
-import com.example.whiplash.domain.user.User;
-import jakarta.persistence.*;
-import lombok.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "search_history")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class SearchHistory {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String keyword;
-    private Integer resultCount;
-    private LocalDateTime searchedAt;
-
-    // 단방향: SearchHistory → User
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-}
