+++ src/main/resources/application.yml
@@ -1,0 +1,21 @@
+spring:
+  datasource:
+    url: jdbc:mysql://localhost:3306/whiplash
+    username: root
+    password: 1111
+    driver-class-name: com.mysql.cj.jdbc.Driver
+  jpa:
+    hibernate:
+      ddl-auto: update
+    properties:
+      hibernate:
+        format_sql: true
+    #        default_batch_fetch_size: 100
+    show-sql: true
+  data:
+    mongodb:
+      host: localhost
+      port: 27017
+      database: database_name
+      username: root
+      password: 1111
